Refactor playground loader to key-based lifecycle
