+++ playground/src/views/core/index.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
-import { computed, ref } from 'vue'
-import type { LabComponentDefinition, PlaygroundPropValue } from '@/library/catalog'
-import { getComponentDefinition } from '@/library/catalog'
+import { computed, ref, shallowRef } from 'vue'
+import type { PlaygroundComponent, PlaygroundPropValue } from '@/library/types'
+import { findCatalogComponent } from '@/library/catalog'
+import { loadPlaygroundDemo } from '@/library/demos'
 import Controls from './controls/index.vue'
 import NotFound from './notFound.vue'
 import Preview from './preview.vue'
@@ -10,13 +11,27 @@ const props = defineProps<{
   componentId: string
 }>()
 
-const definition = computed<LabComponentDefinition | undefined>(() => getComponentDefinition(props.componentId))
-
+const definition = shallowRef<PlaygroundComponent | null>(null)
 const demoProps = ref<Record<string, PlaygroundPropValue>>({})
+const componentKey = computed(() => props.componentId)
+
+const entry = findCatalogComponent(componentKey.value)
+
+if (entry) {
+  const demo = await loadPlaygroundDemo(componentKey.value)
+
+  if (demo) {
+    definition.value = { ...entry, ...demo }
+  }
+}
 </script>
 
 <template>
-  <div v-if="definition" class="grid gap-8 lg:grid-cols-[minmax(0,1.2fr)_22rem]">
+  <div
+    v-if="definition"
+    :key="componentKey"
+    class="grid gap-8 lg:grid-cols-[minmax(0,1.2fr)_22rem]"
+  >
     <Preview
       :key="definition.id"
       :definition="definition"
